+++ dlc_fast_prototyping/src/common/types.rs
@@ -11,7 +11,7 @@ pub type Attestation = SecretKey;
 
 // Other
 pub type Cet = String; // Contract Execution Transaction (esentially not signed Tx)
-pub type ParsedContract<O: Outcome> = Vec<(O, u32)>; // (outcome, payout) pairs.. TODO: nejak to treba vymysliet aby to slo urobit
+pub type ParsedContract<O: Outcome> = Vec<(O, u64)>; // (outcome, payout) pairs.. TODO: nejak to treba vymysliet aby to slo urobit
                                                      // TODO: za predpokladu, ze ParsedContract bude obsahovat len OutcomeU32, tak by to mohol byt iba Vec<u32>
 
 /// The final Bitcoin transaction or any other on-chain transaction type
@@ -179,7 +179,7 @@ pub struct PayoutInterval {
 #[serde(rename_all = "camelCase")]
 pub struct PayoutPoint {
     pub event_outcome: u32,
-    pub outcome_payout: u32, // Amount (btc cargo)
+    pub outcome_payout: u64, // Amount (btc cargo)
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
